% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape.R
\name{get_all_narigs}
\alias{get_all_narigs}
\title{get_all_narigs}
\usage{
get_all_narigs()
}
\value{
A data frame of historical rig counts
}
\description{
Pulls historical rig counts (oil/gas splits and trajectory) from
Baker Hughes
}
